server/requirements.txt
influxdb
python-dotenv
paho-mqtt